requirements.txt
python-telegram-bot
firebase-admin
pyrebase
pyrebase4
pytelegrambotapi
psutil
pygetwindow